306 - Refactor the implementation and move helper functions
