+++ scripts/scripts.js
@@ -28,7 +28,15 @@ import {
   getLocale,
 } from './common.js';
 
-import { isVideoLink, isSoundcloudLink, isLowResolutionVideoUrl, addVideoShowHandler, addSoundcloudShowHandler } from './video-helper.js';
+import {
+  isVideoLink,
+  isSoundcloudLink,
+  isLowResolutionVideoUrl,
+  addVideoShowHandler,
+  addSoundcloudShowHandler,
+  hasVideoOnPage,
+  loadVideoJs,
+} from './video-helper.js';
 
 import { validateCountries } from './validate-countries.js';
 
@@ -863,28 +871,6 @@ function loadDelayed() {
 }
 
 /**
- * Determines whether the page contains any video-related elements.
- * @returns {boolean} True if a video block is present on the page.
- */
-function hasVideoOnPage() {
-  return !!(
-    document.querySelector('.video-js') ||
-    document.querySelector('.link-with-video') ||
-    document.querySelector('.text-link-with-video') ||
-    document.querySelector('.v2-video__big-play-button') ||
-    document.querySelector('.v2-resource-gallery__video-list-item .icon-play-video')
-  );
-}
-
-/**
- * Loads the delayed module and initializes Video.js.
- */
-async function loadVideoSupport() {
-  const delayedModule = await import('./delayed.js');
-  await delayedModule.loadVideoJs();
-}
-
-/**
  * Main page initialization logic.
  * Loads eager/lazy resources and conditionally loads video support.
  */
@@ -893,7 +879,7 @@ async function loadPage() {
   await loadLazy(document);
 
   if (hasVideoOnPage()) {
-    await loadVideoSupport();
+    await loadVideoJs();
   } else {
     loadDelayed();
   }
